Sonar Category C cleanup: dead code + redundant casts (S1905, S1068, S1144, S1130) (#3395)
* Remove unnecessary casts (S1905)

Drop 18 redundant casts. Each was verified safe: every '(long) i * X'
multiplication has a long right operand (X is increment, PTR_SIZE,
ENUM_SERVICE_STATUS_PROCESS_SIZE, ADDRESS.byteSize(), etc.), so the product was
already computed in long and no narrowing/overflow was being prevented. The two
'(int)' casts on short/byte values are redundant because the '&' operator
already promotes to int, and getPartitions() is already declared List<HWPartition>.
No behavior change.

* Remove unused private fields (S1068)

Delete eight genuinely dead private fields: unused loggers (MacGpuStatsFFM,
WindowsFileSystemFFM, DragonFlyBsdOSProcessJNA), write-only fields whose derived
state is what's actually used (MacGpuStatsFFM.cardName, CupsFunctions.CUPS_LIBRARY
- the bound method handles use the local lookup), and leftover constants
(AdlUtilFFM.PERF_GPU_ACTIVITY_OFFSET, WindowsGraphicsCardJNA.MB_TO_BYTES from a
removed LHM code path, OshiMetricsTest.hal).

NOT deleted: DiskMetrics.diskStores. It looks unused but is a deliberate strong
reference (GC root) that keeps the disk objects alive, because Micrometer's
FunctionCounter only holds WeakReferences to them; deleting it would silently
break the disk metrics after a GC. Documented and marked @SuppressWarnings("java:S1068").

* Remove unused private methods and constructors (S1144)

Delete four dead overloads (not prevent-instantiation idioms - each class is
instantiated via its other constructor): the pre-diskType constructors of
MacHWDiskStoreFFM / WindowsHWDiskStoreFFM / WindowsHWDiskStoreJNA (the diskType
variants are the ones getDisks() uses), and LinuxGraphicsCard's
queryLspciMemorySize(String) overload, which nothing calls - the caller already
runs lspci and invokes the List<String> overload directly. No behavior change.

* Remove declaration of unthrown exception (S1130)

MacInstalledApps.readStringValue declared 'throws Exception' but does only string
parsing and cannot throw it. Removing the clause is safe: the sole caller wraps
it in a catch(Exception) that remains valid (it still catches RuntimeException
and other calls in the same try). No behavior change.

The six JMXOshiAgentServer S1130 findings are intentionally left: they are
@Override stubs whose throws clauses mirror the JMX MBeanServer interface
contract (Java forbids an override from adding new checked exceptions), so
removing Sonar's flagged subset would make them inconsistent with their siblings.

+++ oshi-core-ffm/src/main/java/oshi/driver/unix/aix/PsInfoFFM.java
@@ -89,7 +89,7 @@ public final class PsInfoFFM {
                 long argp = bufStart == 0 ? 0 : readPointerFromBuffer(buffer, argv - bufStart, increment);
                 if (argp > 0) {
                     for (int i = 0; i < argc; i++) {
-                        long offset = argp + (long) i * increment;
+                        long offset = argp + i * increment;
                         bufStart = conditionallyReadPage(fd, buffer, bufStart, offset);
                         argPtr[i] = bufStart == 0 ? 0 : readPointerFromBuffer(buffer, offset - bufStart, increment);
                     }
